Refactor and enhance HTTP handling with new classes and updated structures
- Added new classes for HttpHeaders, HttpBody, and HttpURI to improve HTTP request and response management.
- Updated Client class to include local and remote address handling, enhancing client connection management.
- Refactored ConfigParser to expose server configurations through a new accessor method.
- Improved Makefile to include additional source files for better project structure.
- Cleaned up includes across various files for consistency and maintainability.
- Enhanced error handling and logging throughout the HTTP processing system.

+++ srcs/MethodHandlers/GetMethodHandler.cpp
@@ -1,4 +1,4 @@
-#include "GetMethodHandler.hpp"
+#include "../../includes/GetMethodHandler.hpp"
 
 /*
 ** ------------------------------- CONSTRUCTOR --------------------------------
@@ -164,7 +164,7 @@ void GetMethodHandler::serveFile(const std::string &filePath, HttpResponse &resp
 
 	// Check file size (prevent loading huge files into memory)
 	const size_t MAX_FILE_SIZE = 100 * 1024 * 1024; // 100MB limit
-	if (fileStat.st_size > MAX_FILE_SIZE)
+	if (static_cast<size_t>(fileStat.st_size) > MAX_FILE_SIZE)
 	{
 		Logger::log(Logger::ERROR, "File too large: " + filePath);
 		response.setStatus(413, "Payload Too Large");
